task_backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-core==0.1.10
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3